+++ CHANGELOG.md
@@ -28,6 +28,9 @@
   - `bedtools/bedtools_sort`: Sorts a feature file (bed/gff/vcf) by chromosome and other criteria (PR #98).
 
 
+* `agat/agat_convert_genscan2gff`: convert a genscan file into a GFF file (PR #100).
+
+
 ## MINOR CHANGES
 
 * `busco` components: update BUSCO to `5.7.1` (PR #72).